requirements.txt
# AWS Dependencies
awsiotsdk==1.7.1
# Raspberry Pi Dependencies
smbus
getmac
fake_rpi